+++ anvil_consortium_manager/anvil_audit.py
@@ -97,19 +97,19 @@ class AnVILAuditResults(ABC):
         """
         return not self.errors and not self.not_in_app
 
-    def to_json(
+    def export(
         self, include_verified=True, include_errors=True, include_not_in_app=True
     ):
         """Return a dictionary representation of the audit results."""
         x = {}
         if include_verified:
             x["verified"] = [
-                {"id": instance.pk, "instance": str(instance)}
+                {"id": instance.pk, "instance": instance}
                 for instance in self.get_verified()
             ]
         if include_errors:
             x["errors"] = [
-                {"id": k.pk, "instance": str(k), "errors": v}
+                {"id": k.pk, "instance": k, "errors": v}
                 for k, v in self.get_errors().items()
             ]
         if include_not_in_app: