+++ anvil_consortium_manager/management/commands/run_anvil_audit.py
@@ -1,10 +1,36 @@
-import json
+import pprint
 
+import django_tables2 as tables
+from django.contrib.sites.models import Site
 from django.core.mail import send_mail
 from django.core.management.base import BaseCommand
+from django.template.loader import render_to_string
 
 from ... import models
 from ...anvil_api import AnVILAPIError
+
+
+class ErrorsTable(tables.Table):
+    id = tables.Column(orderable=False)
+    instance = tables.Column(
+        orderable=False,
+        linkify=lambda value, table: "https://{domain}{url}".format(
+            domain=table.site.domain, url=value.get_absolute_url()
+        ),
+    )
+    errors = tables.Column(orderable=False)
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        # Set the site here so it only hits the db once.
+        self.site = Site.objects.get_current()
+
+    def render_errors(self, value):
+        return ", ".join(value)
+
+
+class NotInAppTable(tables.Table):
+    instance = tables.Column(orderable=False)
 
 
 class Command(BaseCommand):
@@ -36,7 +62,7 @@ class Command(BaseCommand):
         email = options["email"]
         errors_only = options["errors_only"]
 
-        model_name = model._meta.verbose_name_plural
+        model_name = model._meta.object_name
 
         self.stdout.write("Running on {}... ".format(model_name), ending="")
         try:
@@ -47,27 +73,37 @@ class Command(BaseCommand):
         else:
             if not results.ok():
                 self.stdout.write(self.style.ERROR("problems found."))
-                report = json.dumps(results.to_json(include_verified=False), indent=2)
-                if email:
-                    send_mail(
-                        "AnVIL Audit for {} -- errors".format(model_name),
-                        report,
-                        None,
-                        [email],
-                        fail_silently=False,
-                    )
-                else:
-                    self.stdout.write(report)
+                self.stdout.write(
+                    pprint.pformat(results.export(include_verified=False))
+                )
             else:
-                self.stdout.write(self.style.ERROR("ok!"))
-                if email and not errors_only:
-                    send_mail(
-                        "AnVIL Audit for {} -- ok".format(model_name),
-                        "Audit ok ({} instances)".format(len(results.get_verified())),
-                        None,
-                        [email],
-                        fail_silently=False,
-                    )
+                self.stdout.write(self.style.SUCCESS("ok!"))
+
+            if email and (not errors_only) or (errors_only and not results.ok()):
+                # Set up the email message.
+                subject = "AnVIL audit for {} -- {}".format(
+                    model_name, "ok" if results.ok() else "errors!"
+                )
+                exported_results = results.export()
+                html_body = render_to_string(
+                    "anvil_consortium_manager/email_audit_report.html",
+                    context={
+                        "model_name": model_name,
+                        "verified_results": exported_results["verified"],
+                        "errors_table": ErrorsTable(exported_results["errors"]),
+                        "not_in_app_table": NotInAppTable(
+                            exported_results["not_in_app"]
+                        ),
+                    },
+                )
+                send_mail(
+                    subject,
+                    pprint.pformat(exported_results),
+                    None,
+                    [email],
+                    fail_silently=False,
+                    html_message=html_body,
+                )
 
     def handle(self, *args, **options):