+++ anvil_consortium_manager/tests/test_commands.py
@@ -1,13 +1,16 @@
 """Tests for management commands in `anvil_consortium_manager`."""
 
+import pprint
 from io import StringIO
+from unittest import skip
 
 import responses
+from django.contrib.sites.models import Site
 from django.core import mail
 from django.core.management import CommandError, call_command
-from django.test import TestCase
+from django.test import TestCase, override_settings
 
-from .. import anvil_audit
+from .. import anvil_audit, models
 from . import factories
 from .utils import AnVILAPIMockTestMixin
 
@@ -48,10 +51,10 @@ class RunAnvilAuditTest(AnVILAPIMockTestMixin, TestCase):
         )
         out = StringIO()
         call_command("run_anvil_audit", "--no-color", stdout=out)
-        self.assertIn("billing projects... ok!", out.getvalue())
-        self.assertIn("accounts... ok!", out.getvalue())
-        self.assertIn("managed groups... ok!", out.getvalue())
-        self.assertIn("workspaces... ok!", out.getvalue())
+        self.assertIn("BillingProject... ok!", out.getvalue())
+        self.assertIn("Account... ok!", out.getvalue())
+        self.assertIn("ManagedGroup... ok!", out.getvalue())
+        self.assertIn("Workspace... ok!", out.getvalue())
 
     def test_command_output_multiple_models(self):
         """Can audit multiple models at the same time."""
@@ -62,8 +65,8 @@ class RunAnvilAuditTest(AnVILAPIMockTestMixin, TestCase):
             models=["BillingProject", "Account"],
             stdout=out,
         )
-        self.assertIn("billing projects... ok!", out.getvalue())
-        self.assertIn("accounts... ok!", out.getvalue())
+        self.assertIn("BillingProject... ok!", out.getvalue())
+        self.assertIn("Account... ok!", out.getvalue())
 
     def test_command_output_billing_project_no_instances(self):
         """Test command output."""
@@ -71,13 +74,13 @@ class RunAnvilAuditTest(AnVILAPIMockTestMixin, TestCase):
         call_command(
             "run_anvil_audit", "--no-color", models=["BillingProject"], stdout=out
         )
-        self.assertIn("billing projects... ok!", out.getvalue())
+        self.assertIn("BillingProject... ok!", out.getvalue())
 
     def test_command_output_account_no_instances(self):
         """Test command output."""
         out = StringIO()
         call_command("run_anvil_audit", "--no-color", models=["Account"], stdout=out)
-        self.assertIn("accounts... ok!", out.getvalue())
+        self.assertIn("Account... ok!", out.getvalue())
 
     def test_command_output_managed_group_no_instances(self):
         """Test command output."""
@@ -91,7 +94,7 @@ class RunAnvilAuditTest(AnVILAPIMockTestMixin, TestCase):
         call_command(
             "run_anvil_audit", "--no-color", models=["ManagedGroup"], stdout=out
         )
-        self.assertIn("managed groups... ok!", out.getvalue())
+        self.assertIn("ManagedGroup... ok!", out.getvalue())
 
     def test_command_output_workspace_no_instances(self):
         """Test command output."""
@@ -103,7 +106,7 @@ class RunAnvilAuditTest(AnVILAPIMockTestMixin, TestCase):
         )
         out = StringIO()
         call_command("run_anvil_audit", "--no-color", models=["Workspace"], stdout=out)
-        self.assertIn("workspaces... ok!", out.getvalue())
+        self.assertIn("Workspace... ok!", out.getvalue())
 
     def test_command_run_audit_one_instance_ok(self):
         """Test command output."""
@@ -115,7 +118,7 @@ class RunAnvilAuditTest(AnVILAPIMockTestMixin, TestCase):
         call_command(
             "run_anvil_audit", "--no-color", models=["BillingProject"], stdout=out
         )
-        self.assertIn("billing projects... ok!", out.getvalue())
+        self.assertIn("BillingProject... ok!", out.getvalue())
         self.assertNotIn("errors", out.getvalue())
         self.assertNotIn("not_in_app", out.getvalue())
 
@@ -133,11 +136,17 @@ class RunAnvilAuditTest(AnVILAPIMockTestMixin, TestCase):
             email="test@example.com",
             stdout=out,
         )
-        self.assertIn("billing projects... ok!", out.getvalue())
+        self.assertIn("BillingProject... ok!", out.getvalue())
         # One message has been sent by default.
         self.assertEqual(len(mail.outbox), 1)
-        self.assertIn("ok", mail.outbox[0].subject)
-        self.assertEqual(mail.outbox[0].body, "Audit ok (1 instances)")
+        email = mail.outbox[0]
+        self.assertEqual(email.to, ["test@example.com"])
+        self.assertIn("ok", email.subject)
+        # Text body.
+        audit_results = models.BillingProject.anvil_audit()
+        self.assertEqual(pprint.pformat(audit_results.export()), email.body)
+        # HTML body.
+        self.assertEqual(len(email.alternatives), 1)
 
     def test_command_run_audit_ok_email_errors_only(self):
         """Test command output when email and errors_only is set."""
@@ -154,7 +163,7 @@ class RunAnvilAuditTest(AnVILAPIMockTestMixin, TestCase):
             errors_only=True,
             stdout=out,
         )
-        self.assertIn("billing projects... ok!", out.getvalue())
+        self.assertIn("BillingProject... ok!", out.getvalue())
         # No message has been sent by default.
         self.assertEqual(len(mail.outbox), 0)
 
@@ -170,8 +179,8 @@ class RunAnvilAuditTest(AnVILAPIMockTestMixin, TestCase):
         call_command(
             "run_anvil_audit", "--no-color", models=["BillingProject"], stdout=out
         )
-        self.assertIn("billing projects... problems found.", out.getvalue())
-        self.assertIn(""""errors":""", out.getvalue())
+        self.assertIn("BillingProject... problems found.", out.getvalue())
+        self.assertIn("""'errors':""", out.getvalue())
         self.assertIn(
             anvil_audit.BillingProjectAuditResults.ERROR_NOT_IN_ANVIL, out.getvalue()
         )
@@ -192,19 +201,68 @@ class RunAnvilAuditTest(AnVILAPIMockTestMixin, TestCase):
             email="test@example.com",
             stdout=out,
         )
-        self.assertIn("billing projects... problems found.", out.getvalue())
+        self.assertIn("BillingProject... problems found.", out.getvalue())
         # Not printed to stdout.
-        self.assertNotIn(""""errors":""", out.getvalue())
+        self.assertIn("""'errors':""", out.getvalue())
         # One message has been sent.
         self.assertEqual(len(mail.outbox), 1)
-        self.assertEqual(mail.outbox[0].to, ["test@example.com"])
-        self.assertIn("errors", mail.outbox[0].subject)
-        # Instead in the email body:
-        self.assertIn(""""errors":""", mail.outbox[0].body)
-        self.assertIn(
-            anvil_audit.BillingProjectAuditResults.ERROR_NOT_IN_ANVIL,
-            mail.outbox[0].body,
+        email = mail.outbox[0]
+        self.assertEqual(email.to, ["test@example.com"])
+        self.assertIn("errors", email.subject)
+        # Text body.
+        audit_results = models.BillingProject.anvil_audit()
+        self.assertEqual(pprint.pformat(audit_results.export()), email.body)
+        # HTML body.
+        self.assertEqual(len(email.alternatives), 1)
+
+    def test_command_run_audit_not_ok_email_has_html_link(self):
+        """Test command output when BillingProject audit is not ok with email specified."""
+        billing_project = factories.BillingProjectFactory.create()
+        # Add a response.
+        api_url = self.get_api_url_billing_project(billing_project.name)
+        self.anvil_response_mock.add(
+            responses.GET, api_url, status=404, json={"message": "error"}
         )
+        out = StringIO()
+        call_command(
+            "run_anvil_audit",
+            "--no-color",
+            models=["BillingProject"],
+            email="test@example.com",
+            stdout=out,
+        )
+        email = mail.outbox[0]
+        self.assertEqual(len(email.alternatives), 1)
+        html_fragment = """<a href="https://example.com{url}">{obj}</a>""".format(
+            obj=str(billing_project), url=billing_project.get_absolute_url()
+        )
+        self.assertInHTML(html_fragment, email.alternatives[0][0])
+
+    @override_settings(SITE_ID=2)
+    def test_command_run_audit_not_ok_email_has_html_link_different_domain(self):
+        """Test command output when BillingProject audit is not ok with email specified."""
+        site = Site.objects.create(domain="foobar.com", name="test")
+        site.save()
+        billing_project = factories.BillingProjectFactory.create()
+        # Add a response.
+        api_url = self.get_api_url_billing_project(billing_project.name)
+        self.anvil_response_mock.add(
+            responses.GET, api_url, status=404, json={"message": "error"}
+        )
+        out = StringIO()
+        call_command(
+            "run_anvil_audit",
+            "--no-color",
+            models=["BillingProject"],
+            email="test@example.com",
+            stdout=out,
+        )
+        email = mail.outbox[0]
+        self.assertEqual(len(email.alternatives), 1)
+        html_fragment = """<a href="https://foobar.com{url}">{obj}</a>""".format(
+            obj=str(billing_project), url=billing_project.get_absolute_url()
+        )
+        self.assertInHTML(html_fragment, email.alternatives[0][0])
 
     def test_command_run_audit_api_error(self):
         """Test command output when BillingProject audit is not ok."""
@@ -218,4 +276,22 @@ class RunAnvilAuditTest(AnVILAPIMockTestMixin, TestCase):
         call_command(
             "run_anvil_audit", "--no-color", models=["BillingProject"], stdout=out
         )
-        self.assertIn("billing projects... API error.", out.getvalue())
+        self.assertIn("BillingProject... API error.", out.getvalue())
+
+    # This test is complicated so skipping for now.
+    # When trying to change the settings, the test attempts to repopulate the
+    # workspace registry. This then causes an error. Skip it until we figure out
+    # how best to handle this situation.
+    @skip
+    def test_command_without_sites(self):
+        """Test command behavior without the Sites framework enabled."""
+        out = StringIO()
+        with self.modify_settings(INSTALLED_APPS={"remove": ["django.contrib.sites"]}):
+            #            with self.assertRaises(ImproperlyConfigured):
+            call_command(
+                "run_anvil_audit",
+                "--no-color",
+                models=["BillingProject"],
+                email="test@example.com",
+                stdout=out,
+            )