+++ example_site/settings.py
@@ -71,6 +71,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "django.contrib.admin",
+    "django.contrib.sites",
     "django.forms",
     # Third party apps
     "crispy_forms",